Impelemnt simple view-change
The view-change logic is described in the README. The view-change
feature is disabled by default. We do not test the behaviour of when a
node comes back up after being down, it is not expected to work because
there is no logic that let's the node catch-up to the latest state. We
only make sure that if a leader is down, then a new leader is selected
and is able to continue to create blocks. This is the first of many
steps to fully implement a robust view-change mechanism.